+++ Assets/Plugins/StreamChat/Core/DTO/Requests/ChannelConfigRequestDTO.cs
@@ -33,7 +33,7 @@ namespace StreamChat.Core.DTO.Requests
         /// Overrides max message length
         /// </summary>
         [Newtonsoft.Json.JsonProperty("max_message_length", Required = Newtonsoft.Json.Required.Default, NullValueHandling = Newtonsoft.Json.NullValueHandling.Ignore)]
-        public double? MaxMessageLength { get; set; }
+        public int? MaxMessageLength { get; set; }
 
         /// <summary>
         /// Enables message quotes
